@@ -1,7 +1,6 @@
 from frmodel.base.D2.frame._frame_channel import _Frame2DChannel
-from frmodel.base.D2.frame._frame_kmeans import _Frame2DKmeans
 from frmodel.base.D2.frame._frame_loader import _Frame2DLoader
 from frmodel.base.D2.frame._frame_partition import _Frame2DPartition
 from frmodel.base.D2.frame._frame_scaling import _Frame2DScaling
 
-__all__ = ['_Frame2DScaling', '_Frame2DChannel', '_Frame2DLoader', '_Frame2DPartition', '_Frame2DKmeans']
+__all__ = ['_Frame2DScaling', '_Frame2DChannel', '_Frame2DLoader', '_Frame2DPartition']
