@@ -1,44 +1,23 @@
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
+from abc import ABC
+from typing import TYPE_CHECKING
 
-import numpy as np
 from sklearn.preprocessing import minmax_scale as sk_minmax_scale
 from sklearn.preprocessing import normalize as sk_normalize
 
-from frmodel.base.consts import CONSTS
-
-CHANNEL = CONSTS.CHANNEL
-MAX_RGB = 255
+if TYPE_CHECKING:
+    from frmodel.base.D2.frame2D import Frame2D
 
 class _Frame2DScaling(ABC):
 
-    data: np.ndarray
-    
-    @abstractmethod
-    def data_rgb(self): ...
-    
-    @abstractmethod
-    def data_chn(self, *args, **kwargs): ...
+    def normalize(self: 'Frame2D', **kwargs) -> _Frame2DScaling:
+        return self.scale(sk_normalize, **kwargs)
 
-    @abstractmethod
-    def width(self): ...
+    def minmax_scale(self: 'Frame2D', **kwargs) -> _Frame2DScaling:
+        return self.scale(sk_minmax_scale, **kwargs)
 
-    @abstractmethod
-    def height(self): ...
-    
-    @abstractmethod
-    def slide_xy(self, *args, **kwargs): ...
-
-    # noinspection PyArgumentList
-    @classmethod
-    def init(cls, *args, **kwargs):
-        return cls(*args, **kwargs)
-
-    def normalize(self) -> _Frame2DScaling:
+    def scale(self: 'Frame2D', scaler, **scaler_kwargs):
         shape = self.data.shape
-        return self.init(sk_normalize(self.data.reshape([-1, shape[-1]]), axis=0).reshape(shape))
-
-    def minmax_scale(self) -> _Frame2DScaling:
-        shape = self.data.shape
-        return self.init(sk_minmax_scale(self.data.reshape([-1, shape[-1]]), axis=0).reshape(shape))
+        return self.create(data=scaler(self.data_flatten_xy(), **scaler_kwargs).reshape(shape),
+                           labels=self.labels)
